+++ packages/expo/ios/AppDelegates/ExpoAppDelegate.swift
@@ -147,7 +147,15 @@ open class ExpoAppDelegate: NSObject, @preconcurrency ReactNativeFactoryProvider
   }
 #endif
 
-  // TODO: - Responding to Environment Changes
+  // MARK: - Responding to Environment Changes
+  
+#if os(iOS) || os(tvOS)
+
+  open func applicationDidReceiveMemoryWarning(_ application: UIApplication) {
+    ExpoAppDelegateSubscriberManager.applicationDidReceiveMemoryWarning(application)
+  }
+
+#endif
 
   // TODO: - Managing App State Restoration